Added get_current_average_balance() to PaymentAccount.
Note: it currently uses the calendar month to figure everything out. I need to
change this to be the "payment cycle" which I assume would be based on what day
of the month the account was created, and then every 30 days (not 30 or 31) from
then.

Warning message still needs implementation.